[task #9511] move date_part, date_trunc, date_bin functions to datafusion-functions
