+++ datafusion/proto/src/logical_plan/from_proto.rs
@@ -53,8 +53,8 @@ use datafusion_expr::{
     array_repeat, array_replace, array_replace_all, array_replace_n, array_resize,
     array_slice, array_sort, array_union, arrow_typeof, ascii, asinh, atan, atan2, atanh,
     bit_length, btrim, cbrt, ceil, character_length, chr, coalesce, concat_expr,
-    concat_ws_expr, cos, cosh, cot, current_date, current_time, date_bin, date_part,
-    date_trunc, degrees, digest, ends_with, exp,
+    concat_ws_expr, cos, cosh, cot, current_date, current_time, date_bin, degrees,
+    digest, ends_with, exp,
     expr::{self, InList, Sort, WindowFunction},
     factorial, find_in_set, flatten, floor, from_unixtime, gcd, initcap, iszero, lcm,
     left, levenshtein, ln, log, log10, log2,
@@ -502,8 +502,6 @@ impl From<&protobuf::ScalarFunction> for BuiltinScalarFunction {
             ScalarFunction::ArrayUnion => Self::ArrayUnion,
             ScalarFunction::ArrayResize => Self::ArrayResize,
             ScalarFunction::Array => Self::MakeArray,
-            ScalarFunction::DatePart => Self::DatePart,
-            ScalarFunction::DateTrunc => Self::DateTrunc,
             ScalarFunction::DateBin => Self::DateBin,
             ScalarFunction::Md5 => Self::MD5,
             ScalarFunction::Sha224 => Self::SHA224,
@@ -1581,14 +1579,7 @@ pub fn parse_expr(
                 ScalarFunction::Rtrim => {
                     Ok(rtrim(parse_expr(&args[0], registry, codec)?))
                 }
-                ScalarFunction::DatePart => Ok(date_part(
-                    parse_expr(&args[0], registry, codec)?,
-                    parse_expr(&args[1], registry, codec)?,
-                )),
-                ScalarFunction::DateTrunc => Ok(date_trunc(
-                    parse_expr(&args[0], registry, codec)?,
-                    parse_expr(&args[1], registry, codec)?,
-                )),
+
                 ScalarFunction::DateBin => Ok(date_bin(
                     parse_expr(&args[0], registry, codec)?,
                     parse_expr(&args[1], registry, codec)?,
