+++ datafusion/sql/Cargo.toml
@@ -41,6 +41,7 @@ arrow = { workspace = true }
 arrow-schema = { workspace = true }
 datafusion-common = { workspace = true, default-features = true }
 datafusion-expr = { workspace = true }
+datafusion-functions = { workspace = true }
 log = { workspace = true }
 sqlparser = { workspace = true }
